Cleanup and javadoc
