+++ src/main/java/com/minecolonies/core/datalistener/util/ModRemovalorder.java
@@ -1,0 +1,17 @@
+package com.minecolonies.core.datalistener.util;
+
+import net.minecraft.resources.ResourceLocation;
+
+/**
+ * Removal order for removing all datapack entries from an entire mod namespace.
+ *
+ * @param modId the given mod id.
+ */
+public record ModRemovalorder(String modId) implements RemovalOrder
+{
+    @Override
+    public boolean test(final ResourceLocation resourceLocation)
+    {
+        return resourceLocation.getNamespace().equals(modId);
+    }
+}
