+++ src/main/java/com/minecolonies/core/datalistener/util/RemovalOrder.java
@@ -1,0 +1,12 @@
+package com.minecolonies.core.datalistener.util;
+
+import net.minecraft.resources.ResourceLocation;
+
+import java.util.function.Predicate;
+
+/**
+ * Interface for implementing different removal orders.
+ */
+public interface RemovalOrder extends Predicate<ResourceLocation>
+{
+}
